analyzer: fix searching/resolving binded symbols
Binded symbols such as C.Foo or JS.Foo don't have module_name prepended so they need to be searched in every entry in the store's symbols. To find the symbol in an efficient manner, the location is stored inside the binded_symbol_locations array and that array is then used for getting the path of that said symbol name.