analyzer: implement semantic analyzer test (#291)
This PR implements a similar system to #269 but for semantic analysis. This also transfers all the "test-only" methods to a new `analyzer.an_test_utils` submodule.

From here on, changes made from #159 will be gradually integrated as separate PRs with own test units apart from the ones in the `v.checker` test. 

* todo

* analyzer: rename analyzer.v to semantic_analyzer.v

* analyzer: transfer test-only methods to an_test_utils module

* analyzer: update semantic analyzer API

* anayzer, test_utils: fix tests

* analyzer/test_files/semantic_analyzer: add import_symbols test

* Merge branch 'use-tree-sitter' of https://github.com/vlang/vls into analyzer-implement-semantic-analyzer-test

* analyzer: fmt

* test_utils: fix test_file_parser_test