cmd/vls, server: implement auto-shutdown mechanism (#294)
This PR implements an auto-shutdown mechanism in which after a certain period of time, a `shutdown` request will be internally executed to gracefully kill the process. This mechanism is used when clients such as VSCode do not properly shutdown the language server processes when it is abruptly exit.

The system is not yet perfect. Still needs an additional heuristic that needs to check to client's PID to avoid false positives (like opening the editor idle for N minutes) but it will be in a separate time.

Closes #293. Also thanks @hungrybluedev  for the heads up

* todo

* server, cmd/vls: transfer connection test to cmd/vls

* vls: revert build options to -g

* cmd/vls: add timeout flag

* cmd/vls, server: add shutdown timeout mechanism

* server: make shutdown executed regardless of the server status

* cmd/vls: disable timeout on child process

let parent vls host handle and execute the shutdown to the child process

* cmd/vls: add timeout test to connection_test

* cmd/vls: add comment code for upcoming vls internal request handling

* cmd/vls, server: fmt