analyzer, server, tree_sitter/_v: add static typings to tree-sitter nodes (#341)
This PR adds static typing support to `tree_sitter` nodes. In order to add static type support for tree-sitter nodes, a set of generic types from `Parser<T>` to `TreeCursor<T>` is introduced in order to wrap over existing C types and pass in node type name converter or the `NodeTypeFactory<T>` interface in order to have a method or function that will convert string'ed node types into concrete types.

As to utilizing it, a [section](https://tree-sitter.github.io/tree-sitter/using-parsers#static-node-types) in the Tree-Sitter docs describes how the grammar's `node-types.json` can be used to add static type analysis to TS nodes. In the case of `tree_sitter_v`, the `node-types.json` is utilized by feeding into a `vsh` shell program that will generate `.v` file that contains an enum named `NodeType` which contains all the node types and a `NodeTypeFactory` implementation that will convert the string types into respective `NodeType` member. (For example, if we have a node whose type is `"call_expression"` then static type equiv is `NodeType.call_expression`)

By adding static types to tree-sitter nodes, the compiler will be able to catch most of the errors early on making it easier to maintain in the long run.