analyzer, server, tree_sitter: implement common interface for text buffers (#354)
This is a follow up by #353. This PR allows any text source (string, []rune, and etc) to be used by tree-sitter nodes when getting a specific portion of the text as long as it implements methods as described from the`SourceText` interface.

This will also minimize code changes in the future whenever there are modifications to the text buffer type to be used same with what happened in PR #353. 

This also reverts the previously renamed `code` method back to `text`.

* ropes: modify substr, add test

* tree_sitter: implement SourceText interface, replace code/text method

* ropes: fix at out of bounds panic

* server: migrate to tree_sitter.SourceText interface

* analyzer: migrate to tree_sitter.SourceText

* analyzer: fix tests

* server/tests: fix positions_test.v