cmd/vls: fix child stderr output monitoring (#364)
Fixes #360. This PR fixes the increased CPU usage of host process on Windows caused by a "less used" select loop and a continuous logging of stderr regardless of its content. The offending code has been removed and `stderr_read` is replaced with `stderr_flush` so that it would only be logged if whenever it has content or is in EOF. 

Before the patch, CPU usage of the host process on standby ranges from 20% up to 45% compared to 0.5% to 6% after the patch.

The change shouldn't regress on platforms other than Windows.

* cmd/vls: remove receive_data

* cmd/vls: use stderr_slurp instead of stderr_read

* cmd/vls: simplify handle_exit if logic