analyzer: replace cur_file_{path, dir, name}, with context system (#372)
Currently the analyzer uses the `cur_file_{path, dir, name}` system in order to access information specific to the particular file and directory. While this is good in the beginning, it starts to experience problems especially when switching files. 

Through the new `Context` struct, we are able to access the same specific information but without the pitfalls of not knowing whether or not the current file info in the store is set in to the right file. 

Nevertheless, each LSP request provides the location of the file so it should be good. Another good thing is that we were able to simplify the inputs and add more information that are often used such as trees.

This also somewhat fixes issue #229 but if and only if it is in child mode (aka without crash reporting) 

This also somewhat fixes issue #229 but if and only if it is in child mode (aka without crash reporting) 

* analyzer, server: remove cur_file_path

* analyzer: use context pattern

* analyzer: fix tests

* analyzer: fix tests

* analyzer: fix tests

* analyzer: use imp.context.file_{name,dir}

* analyzer: fix semantic_analyzer.v

* server: fix text_synchronization.v

* server: fix completion.v

* analyzer: use store.default_context for builtin.v